--- src/model.ts
@@ -1,56 +1,0 @@
-export interface User {
-  userId: number
-  username: string,
-  firstName: string,
-  lastName: string,
-  //dont change authorities to permissions, shit dont work then for some reason
-  authorities: Permission[]
-}
-export class User{
-  userId: number
-  username: string;
-  firstName: string;
-  lastName: string;
-
-  constructor(
-    userId: number,
-    username: string,
-    firstName: string,
-    lastName: string,
-
-  ) {
-    this.userId = userId
-    this.username = username
-    this.firstName = firstName
-    this.lastName = lastName
-  }
-}
-
-export interface Users {
-  users: User[];
-}
-
-export class Permission{
-  name: string;
-  constructor(name: string) {
-    this.name = name
-  }
-}
-
-export interface KontnaGrupa {
-  kontnaGrupaId: number
-  brojKonta: string
-  //nekad se u response dobije kao naziv nekad kao nazivKonta #spaghettiAndMeatballs
-  naziv: string
-  nazivKonta: string
-}
-
-export interface readKontoResponse {
-  content: KontnaGrupa[]
-}
-
-export interface LoginResponse {
-  jwt: string
-}
-
-
